migrate: home page dashboard from ipc handlers to api routes
