--- src/main/ipcHandlers/dashboardHandlers/index.ts
@@ -1,11 +1,0 @@
-import { getChartMetrics } from "./getChartMetrics";
-import { getMetricsSummary } from "./getMetricsSummary";
-import { getRecentTransactions } from "./getRecentTransactions";
-import { getTopProducts } from "./getTopProducts";
-
-export function dashboardHandlers() {
-  getMetricsSummary();
-  getChartMetrics();
-  getRecentTransactions();
-  getTopProducts();
-}
