+++ src/main/modules/dashboard/dashboard.types.ts
@@ -1,0 +1,6 @@
+export interface DashboardDates {
+  startofToday: string;
+  endofToday: string;
+  startofYesterday: string;
+  endofYesterday: string;
+}
