+++ src/main/modules/dashboard/dashboard.utils.ts
@@ -1,0 +1,66 @@
+export const dayNames = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
+
+export const monthNames = [
+  "Jan",
+  "Feb",
+  "Mar",
+  "Apr",
+  "May",
+  "Jun",
+  "Jul",
+  "Aug",
+  "Sep",
+  "Oct",
+  "Nov",
+  "Dec"
+];
+
+export const getDashboardDates = () => {
+  const today = new Date();
+  const yesterday = new Date(today);
+  today.setHours(0, 0, 0, 0);
+  yesterday.setDate(today.getDate() - 1);
+  yesterday.setHours(0, 0, 0, 0);
+
+  const startofToday = today.toISOString();
+  const endofToday = new Date(today.getTime() + 24 * 60 * 60 * 1000 - 1).toISOString();
+  const startofYesterday = yesterday.toISOString();
+  const endofYesterday = new Date(yesterday.getTime() + 24 * 60 * 60 * 1000 - 1).toISOString();
+
+  return {
+    startofToday,
+    endofToday,
+    startofYesterday,
+    endofYesterday
+  };
+};
+
+export const getThisWeekRange = () => {
+  const now = new Date();
+  const startDate = new Date(now);
+  startDate.setDate(now.getDate() - now.getDay());
+  startDate.setHours(0, 0, 0, 0);
+
+  const endDate = new Date(startDate);
+  endDate.setDate(startDate.getDate() + 6);
+  endDate.setHours(23, 59, 59, 999);
+
+  return {
+    startDate: startDate.toISOString(),
+    endDate: endDate.toISOString()
+  };
+};
+
+export const getLast7DaysRange = () => {
+  const startDate = new Date();
+  startDate.setDate(startDate.getDate() - 7);
+  startDate.setHours(0, 0, 0, 0);
+
+  const endDate = new Date();
+  endDate.setHours(23, 59, 59, 999);
+
+  return {
+    startDate: startDate.toISOString(),
+    endDate: endDate.toISOString()
+  };
+};
