+++ src/main/server.ts
@@ -2,6 +2,7 @@ import { serve } from "@hono/node-server";
 import { Hono } from "hono";
 import { cors } from "hono/cors";
 import { customersController } from "./modules/customers/customers.controller";
+import { dashboardController } from "./modules/dashboard/dashboard.controller";
 import { estimatesController } from "./modules/estimates/estimates.controller";
 import { productsController } from "./modules/products/products.controller";
 import { salesController } from "./modules/sales/sales.controller";
@@ -9,6 +10,7 @@ import { salesController } from "./modules/sales/sales.controller";
 const app = new Hono();
 
 app.use("/*", cors());
+app.route("/api/dashboard", dashboardController);
 app.route("/api/products", productsController);
 app.route("/api/customers", customersController);
 app.route("/api/sales", salesController);
