+++ src/main/setupIpcHandlers.ts
@@ -1,7 +1,5 @@
-import { dashboardHandlers } from "./ipcHandlers/dashboardHandlers";
 import { shareHandlers } from "./ipcHandlers/shareHandlers";
 
 export function setupIpcHandlers() {
-  dashboardHandlers();
   shareHandlers();
 }
