+++ src/preload/index.ts
@@ -1,12 +1,5 @@
 import { contextBridge, ipcRenderer } from "electron";
-import type { DashboardApi, ShareApi, TransactionType } from "../shared/types";
-
-const dashboardApi: DashboardApi = {
-  getMetricsSummary: () => ipcRenderer.invoke("dashboardApi:getMetricsSummary"),
-  getChartMetrics: (timePeriod) => ipcRenderer.invoke("dashboardApi:getChartMetrics", timePeriod),
-  getRecentTransactions: (type) => ipcRenderer.invoke("dashboardApi:getRecentTransactions", type),
-  getTopProducts: () => ipcRenderer.invoke("dashboardApi:getTopProducts")
-};
+import type { ShareApi, TransactionType } from "../shared/types";
 
 const shareApi: ShareApi = {
   saveAsPDF: (transactionId: string, type: TransactionType) =>
@@ -18,7 +11,6 @@ if (process.contextIsolated) {
     contextBridge.exposeInMainWorld("electronAPI", {
       printReceipt: (html: string) => ipcRenderer.send("print-receipt", html)
     });
-    contextBridge.exposeInMainWorld("dashboardApi", dashboardApi);
     contextBridge.exposeInMainWorld("shareApi", shareApi);
   } catch (error) {
     console.error(error);
@@ -26,8 +18,6 @@ if (process.contextIsolated) {
 } else {
   // @ts-ignore (define in dts)
   window.electron = electronAPI;
-  // @ts-ignore (define in dts)
-  window.dashboardApi = dashboardApi;
   // @ts-ignore (define in ts)
   window.shareApi = shareApi;
 }
