+++ src/renderer/src/hooks/dashboard/useDashboard.ts
@@ -44,11 +44,17 @@ const handleDelete = async ({ type, id }: MutationVariables) => {
 const handleConvert = async ({ type, id }: MutationVariables) => {
   try {
     if (type === TRANSACTION_TYPE.SALE) {
-      const response = await window.salesApi.convertSaletoEstimate(id);
-      return response;
+      const response = await fetch(`http://localhost:3000/api/sales/${id}/convert`, {
+        method: "POST"
+      });
+      const data = await response.json();
+      return data;
     } else if (type === TRANSACTION_TYPE.ESTIMATE) {
-      const response = await window.estimatesApi.convertEstimateToSale(id);
-      return response;
+      const response = await fetch(`http://localhost:3000/api/estimates/${id}/convert`, {
+        method: "POST"
+      });
+      const data = await response.json();
+      return data;
     } else {
       throw new Error("Something went wrong");
     }
